% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{clean_gps_points}
\alias{clean_gps_points}
\title{Cleans data from Google Sheets to make GPS coordinates consistent.}
\usage{
clean_gps_points(the_data)
}
\arguments{
\item{the_data}{}
}
\value{
Cleaned \code{data.frame}
}
\description{
Cleans data from Google Sheets to make GPS coordinates consistent.
}
\examples{
clean_gps_points(getdata())
}
