% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{clean_site_name_rep_detail}
\alias{clean_site_name_rep_detail}
\title{Clean the site name details in case we don't want "rep 1" or "rep 2" included}
\usage{
clean_site_name_rep_detail(the_data)
}
\arguments{
\item{the_data}{}
}
\value{
Cleaned \code{data.frame}
}
\description{
Clean the site name details in case we don't want "rep 1" or "rep 2" included
}
