% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{do_gs4_auth}
\alias{do_gs4_auth}
\title{Does Google Sheets authentication using a Google Service Account. It looks
for a \code{.json} service account key in the ".secrets" directory.}
\usage{
do_gs4_auth()
}
\value{
Provides authentication analagous to gs4_auth()
}
\description{
Does Google Sheets authentication using a Google Service Account. It looks
for a \code{.json} service account key in the ".secrets" directory.
}
\examples{
# Run before read_sheet() functions
do_gs4_auth()
}
