% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_browseable_seq_data}
\alias{get_browseable_seq_data}
\title{Produce data in a nice clean format for browsing & downloading}
\usage{
get_browseable_seq_data()
}
\value{
a \code{data.frame}
}
\description{
Produce data in a nice clean format for browsing & downloading
}
\examples{
get_browseable_seq_data()
}
