% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_browseable_soil_testing_data}
\alias{get_browseable_soil_testing_data}
\title{Produce soil testing data in a nice clean format for browsing on the app.}
\usage{
get_browseable_soil_testing_data()
}
\value{
a \code{data.frame}
}
\description{
Produce soil testing data in a nice clean format for browsing on the app.
}
\examples{
get_browseable_soil_testing_data()
}
