% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_dna_conc_data}
\alias{get_dna_conc_data}
\title{Produce DNA concentration data in a clean format for browsing on the app.}
\usage{
get_dna_conc_data()
}
\value{
a \code{data.frame}
}
\description{
Produce DNA concentration data in a clean format for browsing on the app.
}
\examples{
get_dna_conc_data()
}
