% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_soil_data}
\alias{get_soil_data}
\title{Get soil spatial data so that soil type areas can be plotted on a
map. These are analagous to GIS polygons. This function does some checking
to see that the \code{soil_type_data.rds} file is younger than other data files.}
\usage{
get_soil_data()
}
\value{
an object of class SpatialPolygonsDataFrame
}
\description{
Get soil spatial data so that soil type areas can be plotted on a
map. These are analagous to GIS polygons. This function does some checking
to see that the \code{soil_type_data.rds} file is younger than other data files.
}
\examples{
get_soil_data()
}
