% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shiny_server}
\alias{shiny_server}
\title{Shiny App, server side}
\usage{
shiny_server(input, output, session)
}
\arguments{
\item{input}{inputs from UI}

\item{output}{outputs to UI}

\item{session}{the individual app session}
}
\description{
Shiny App, server side
}
