% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_soil_data}
\alias{get_soil_data}
\title{Get soil spatial data so that soil type areas can be plotted on a map. These
are analagous to GIS polygons. Because loading up the different polygons takes
a while, this function checks that the \code{soil_type_data.rds} file is younger than
the script that created it (aka this one). If it is younger, there is no
need to remake the file. It can simply be read in from \code{.rds}.}
\usage{
get_soil_data()
}
\value{
an object of class SpatialPolygonsDataFrame
}
\description{
Get soil spatial data so that soil type areas can be plotted on a map. These
are analagous to GIS polygons. Because loading up the different polygons takes
a while, this function checks that the \code{soil_type_data.rds} file is younger than
the script that created it (aka this one). If it is younger, there is no
need to remake the file. It can simply be read in from \code{.rds}.
}
\examples{
get_soil_data()
}
