% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Loads soil data from snapshot with local caching.}
\usage{
getdata(dataset = "sites", snapshot = "20260310")
}
\arguments{
\item{measure}{Which dataset? Can be \code{sites}, \code{soil}, or \code{seq} data.}
}
\value{
Soil data \code{data.frame}
}
\description{
Loads soil data from snapshot with local caching.
}
\examples{
getdata()
getdata(dataset = "soil")
}
