% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Gets relevant data from Google Sheets.}
\usage{
getdata()
}
\value{
a list containing GPS points and site names relevant for mapping. GPS
points are class \code{data.frame}. Site names are class \code{character}.
}
\description{
Gets relevant data from Google Sheets.
}
\examples{
getdata()$points
getdata()$sitenames
}
