% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Loads soil data from snapshot with local caching.}
\usage{
getdata(
  snapshot_path = "data/snapshots/BioDIGS_20250514.csv",
  cache_path = "soil_data.csv"
)
}
\arguments{
\item{snapshot_path}{Path to snapshot file}

\item{cache_path}{Where to store cached data (changing not recommended)}
}
\value{
Soil data \code{data.frame}
}
\description{
Loads soil data from snapshot with local caching.
}
\examples{
getdata()
getdata(snapshot_path = "data/snapshots/BioDIGS_20250514.csv")
}
