% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{retrieve_plot_data}
\alias{retrieve_plot_data}
\title{Partition out the data into a list for quick use in plots.}
\usage{
retrieve_plot_data()
}
\value{
a \code{list} of gps coordinates, site names, and image urls
}
\description{
Partition out the data into a list for quick use in plots.
}
\examples{
# Get gps points as a dataframe:
retrieve_plot_data()$points

# Get site names as a vector:
retrieve_plot_data()$sitenames

# Get urls for images as a vector:
retrieve_plot_data()$image_urls
}
