% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{scrubsitedata}
\alias{scrubsitedata}
\title{Scrubs SITE data according to faculty requests. This is an interactive process that
could likely be automated using a GitHub Action Google Sheet pull at some point.}
\usage{
scrubsitedata(infile, outfile)
}
\arguments{
\item{infile}{Character string specifying the path to the input CSV file containing raw data}

\item{outfile}{Character string specifying the path where the scrubbed data will be saved}
}
\value{
A \code{data.frame} containing the processed data with:
\itemize{
\item Filtered rows based on public sharing permissions
\item Masked GPS measurements
\item Preserved data structure and column names
}
}
\description{
Scrubs SITE data according to faculty requests. This is an interactive process that
could likely be automated using a GitHub Action Google Sheet pull at some point.
}
\examples{
scrubsitedata(infile = "BioDIGS Sample Data and Kit Request MASTER - ALL_SAMPLES - SITE DATA.csv", outfile = "data/snapshots/BioDIGS_sites_20250616.csv")
}
