% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{scrubsoildata}
\alias{scrubsoildata}
\title{Scrubs SOIL data according to faculty requests. This is an interactive process that
could likely be automated using a GitHub Action Google Sheet pull at some point.}
\usage{
scrubsoildata(infile, outfile)
}
\arguments{
\item{infile}{Character string specifying the path to the input CSV file containing raw data}

\item{outfile}{Character string specifying the path where the scrubbed data will be saved}
}
\value{
A \code{data.frame} containing the processed data with:
\itemize{
\item Filtered rows based on public sharing permissions
\item Masked soil property measurements
\item Preserved data structure and column names
}
}
\description{
Scrubs SOIL data according to faculty requests. This is an interactive process that
could likely be automated using a GitHub Action Google Sheet pull at some point.
}
\examples{
scrubsoildata(infile = "BioDIGS Sample Data and Kit Request MASTER - ALL_SAMPLES - SOIL SAMPLES.csv", outfile = "data/snapshots/BioDIGS_soil_20250616.csv")
}
